Remove duplicate methods in AddPublicationUtility.js file and when calling api to add publication or add/update new g2p record, include mechanism_evidence in request json only if mechanism support is set to evidence
